refactor: Remove unused GraphEvents and improve stream event handling in AnthropicClient and OpenAIClient
