+++ api/app/clients/AnthropicClient.js
@@ -9,7 +9,7 @@ const {
   getResponseSender,
   validateVisionModel,
 } = require('librechat-data-provider');
-const { SplitStreamHandler: _Handler, GraphEvents } = require('@librechat/agents');
+const { SplitStreamHandler: _Handler } = require('@librechat/agents');
 const {
   truncateText,
   formatMessage,
@@ -26,11 +26,11 @@ const {
 const { getModelMaxTokens, getModelMaxOutputTokens, matchModelName } = require('~/utils');
 const { spendTokens, spendStructuredTokens } = require('~/models/spendTokens');
 const { encodeAndFormat } = require('~/server/services/Files/images/encode');
+const { createFetch, createStreamEventHandlers } = require('./generators');
 const Tokenizer = require('~/server/services/Tokenizer');
-const { logger, sendEvent } = require('~/config');
 const { sleep } = require('~/server/utils');
 const BaseClient = require('./BaseClient');
-const { createFetch } = require('./fetch');
+const { logger } = require('~/config');
 
 const HUMAN_PROMPT = '\n\nHuman:';
 const AI_PROMPT = '\n\nAssistant:';
@@ -799,30 +799,11 @@ class AnthropicClient extends BaseClient {
     }
 
     logger.debug('[AnthropicClient]', { ...requestOptions });
-    const resRef = new WeakRef(this.options.res);
+    const handlers = createStreamEventHandlers(this.options.res);
     this.streamHandler = new SplitStreamHandler({
       accumulate: true,
       runId: this.responseMessageId,
-      handlers: {
-        [GraphEvents.ON_RUN_STEP]: (event) => {
-          const res = resRef.deref();
-          if (res) {
-            sendEvent(res, event);
-          }
-        },
-        [GraphEvents.ON_MESSAGE_DELTA]: (event) => {
-          const res = resRef.deref();
-          if (res) {
-            sendEvent(res, event);
-          }
-        },
-        [GraphEvents.ON_REASONING_DELTA]: (event) => {
-          const res = resRef.deref();
-          if (res) {
-            sendEvent(res, event);
-          }
-        },
-      },
+      handlers,
     });
 
     let intermediateReply = this.streamHandler.tokens;
